+++ crate/cli/src/actions/mod.rs
@@ -1,4 +1,5 @@
 pub mod access;
+pub mod attributes;
 pub mod certificates;
 pub mod console;
 #[cfg(not(feature = "fips"))]