+++ crate/cli/src/actions/shared/mod.rs
@@ -1,21 +1,15 @@
-mod delete_attributes;
 mod export_key;
-mod get_attributes;
 pub(crate) mod import_key;
 mod locate;
-mod set_attributes;
 pub mod utils;
 
 mod wrap_key;
 
 mod unwrap_key;
 
-pub use delete_attributes::DeleteAttributesAction;
 pub use export_key::{ExportKeyAction, ExportKeyFormat};
-pub use get_attributes::GetAttributesAction;
 pub use import_key::ImportKeyAction;
 pub use locate::LocateObjectsAction;
-pub use set_attributes::{SetAttributesAction, SetOrDeleteAttributes, VendorAttributeCli};
 pub use unwrap_key::UnwrapKeyAction;
 pub use wrap_key::WrapKeyAction;