+++ crate/cli/src/tests/attributes/mod.rs
@@ -1,0 +1,6 @@
+mod delete;
+mod get;
+mod set;
+mod test_set_attribute;
+
+pub(crate) use get::get_attributes;