+++ crate/cli/src/tests/shared/mod.rs
@@ -1,22 +1,16 @@
-pub(crate) use delete_attributes::delete_attributes;
 pub(crate) use destroy::destroy;
 pub(crate) use export::{export_key, ExportKeyParams};
-pub(crate) use get_attributes::get_attributes;
 pub(crate) use import::{import_key, ImportKeyParams};
 pub(crate) use revoke::revoke;
-pub(crate) use set_attributes::set_attributes;
 
-mod delete_attributes;
 mod destroy;
 mod export;
 mod export_import;
-mod get_attributes;
 mod import;
 mod import_export_encodings;
 #[cfg(not(feature = "fips"))]
 mod import_export_wrapping;
 mod locate;
 mod revoke;
-mod set_attributes;
 #[cfg(not(feature = "fips"))]
 mod wrap_unwrap;